betrag in RechnungsPosten ist jetzt int

+++ src/RechnungsPosten.java
@@ -1,7 +1,7 @@
 public abstract class RechnungsPosten {
 	private long zeit;
 	protected int menge;
-	protected double betrag;
+	protected int betrag;
 	protected Artikel artikel;
 	
 	public abstract String toString();
